backend/requirements.txt
# requirements.txt

# Django Core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8

# Face Recognition
insightface==0.7.3
onnxruntime==1.18.0

# Note: If you have an NVIDIA GPU, replace onnxruntime with:
# onnxruntime-gpu==1.18.0

# Vector Search
faiss-cpu==1.8.0

# Note: If you have an NVIDIA GPU, replace faiss-cpu with:
# faiss-gpu==1.7.4

# Image Processing
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# Utilities
python-dateutil==2.9.0
django-widget-tweaks==1.5.0